move request helpers to module and include in spec_helper add inital stab at "factory" for creating and logging in users
