+++ spec/spec_helper.rb
@@ -18,36 +18,15 @@ ActiveRecord::Migration.verbose = false
 ActiveRecord::Migrator.up "db/migrate"
 
 # in case migrations changed what we're testing
-[ User, Cipher, Device, Folder ].each do |c|
+[ Attachment, User, Cipher, Device, Folder ].each do |c|
   c.send(:reset_column_information)
 end
 
 include Rack::Test::Methods
 
-def last_json_response
-  JSON.parse(last_response.body)
-end
+Dir[Rubywarden::App.settings.root + '/spec/support/**/*.rb'].sort.each { |f| require f }
+include Rubywarden::Test::RequestHelpers
 
-def get_json(path, params = {}, headers = {})
-  json_request :get, path, params, headers
-end
-
-def post_json(path, params = {}, headers = {})
-  json_request :post, path, params, headers
-end
-
-def put_json(path, params = {}, headers = {})
-  json_request :put, path, params, headers
-end
-
-def delete_json(path, params = {}, headers = {})
-  json_request :delete, path, params, headers
-end
-
-def json_request(verb, path, params = {}, headers = {})
-  send verb, path, params.to_json,
-    headers.merge({ "CONTENT_TYPE" => "application/json" })
-end
 
 def app
   Rubywarden::App
